refactor: centralize shared logic
