+++ DOCS/ARCHITECTURE.md
@@ -5,9 +5,10 @@ This tool processes weekly "This Week in Rust" Markdown files and prepares messa
 ## Structure
 - **src/main.rs**: Entry point that calls into the CLI module.
 - **src/cli.rs**: Parses command-line arguments and orchestrates the workflow.
-- **src/generator.rs**: Builds Telegram posts and sends them to the API.
-- **src/parser.rs**: Splits Markdown into sections using `pulldown-cmark`.
-- **src/validator.rs**: Checks that generated posts follow Telegram Markdown rules.
+- **src/shared/**: Shared implementation of the generator, parser and validator.
+- **src/generator.rs**: Re-exports the generator functions.
+- **src/parser.rs**: Re-exports the parser implementation.
+- **src/validator.rs**: Re-exports the validator logic.
 - **Cargo.toml**: Defines dependencies such as `pulldown-cmark` and `teloxide`.
 - **last_sent.txt**: Records the last processed issue for the workflow.
 - **src/bin/verify_posts.rs**: Sends posts to Telegram and confirms that they
