--- build_modules/generator.rs
@@ -1,5 +1,0 @@
-#![allow(dead_code)]
-#[path = "../src/generator_shared.rs"]
-mod generator_shared;
-
-pub use generator_shared::*;
