+++ src/generator.rs
@@ -1,4 +1,1 @@
-#[path = "generator_shared.rs"]
-mod generator_shared;
-
-pub use generator_shared::*;
+pub use crate::shared::generator_shared::*;
