+++ src/lib.rs
@@ -2,4 +2,5 @@ pub mod cli;
 pub mod generator;
 pub mod parser;
 pub mod posts;
+pub mod shared;
 pub mod validator;
