+++ src/shared/mod.rs
@@ -1,0 +1,3 @@
+pub mod generator_shared;
+pub mod parser;
+pub mod validator;
